Adding DriverVersion to the DeviceConfig for rolling upgrade support
This PR adds the DriverVersion field to the Spec of the DeviceConfig.
This field is optional, and will be used for rolling upgrade support.
In case the field is empty, then the change to the Spec will cause
the usual upgrade scenario. In case the field is not empty, then any update
to the Spec should also involve updating the DriverVersion field.
DriverVersion is a string that can include any value. The change to the
field will cause the operator to execute rolling upgrade on the cluster