Adjust the KMM module and operator itself to the rolling upgrade flow
As part of the rolling upgrade flow, each node that is going to be
upgrade will be tainted with specific taint. That will cause the
workloads using the kernel module to be removed from the node, so that
the kernel module itself can be upgraded.
This PR adds the toleration for that taint to the KMM module, so that
KMM's pods can continue to run on the upgraded nodes. In addition, the
same toleration is added to the operator's pod, so that it can continue
running in case it's node is being upgraded.