Fixing upgraded node selection in upgrade package
Currently, the upgraded node is being picked up based on the
module-version and module-version-ready labels being present on the node
and containing the same value. But in case there are more then 1 nodes
containing GPU, and at least 2 nodes has finished their upgrade flow,
the first node will always be picked up (since both have label condition
true). This means that the second node (and all the rest) will never be
picked up as upgraded node, and the taint won't be removed from them.
The fix is to look for labels condition AND for the taint being present
on the node.